[FEATURE] Search when the user types  * Refactors product actions file  * fixes tests after refactor  * adds debounce time as a parameter on the payload  * debounces for 500 ms when typing, 0ms when form submits
