+++ src/redux/epics/searchProducts.epic.js
@@ -1,12 +1,15 @@
-import { GET_PRODUCT_RESULTS, sendProductResults } from "../actions"
+import { PRODUCT_ACTIONS, sendProductResults } from "../actions/product.actions"
 import { LCBO_API_KEY, LCBO_BASE_URL } from "../constants"
+import { Observable } from "rxjs"
 
 export const searchProducts = (action$, _, { ajax }) =>
   action$
-    .ofType(GET_PRODUCT_RESULTS)
+    .ofType(PRODUCT_ACTIONS.GET_PRODUCT_RESULTS)
+    .debounce(action => Observable.timer(action.payload.timeInterval))
     .mergeMap(action =>
       ajax(
-        `${LCBO_BASE_URL}products?q=${action.payload}&access_key=${LCBO_API_KEY}`
+        `${LCBO_BASE_URL}products?q=${action.payload
+          .searchString}&access_key=${LCBO_API_KEY}`
       )
     )
     .map(({ response }) => {
