+++ src/redux/reducers/product.reducer.js
@@ -1,18 +1,12 @@
 import { PRODUCT_ACTIONS } from "../actions/product.actions"
-import { SEND_PRODUCT_RESULTS } from "../actions/"
 export const DEFAULT_STATE = {
   wishList: [],
   productList: []
 }
 
-// if (localStorage.getItem("state") !== null) {
-//   //set this to DEFAULT_STATE when we remove fakeProps
-//   fakeProps.wishList = JSON.parse(localStorage.getItem("state")).wishList;
-// }
-
 export const productReducer = (state = DEFAULT_STATE, action) => {
   switch (action.type) {
-    case SEND_PRODUCT_RESULTS:
+    case PRODUCT_ACTIONS.SEND_PRODUCT_RESULTS:
       return {
         ...state,
         productList: action.payload
