+++ src/tests/epics/epics.test.js
@@ -4,7 +4,10 @@ import { Observable } from "rxjs/Observable"
 import "rxjs/add/observable/of"
 
 import { searchProducts } from "../../redux/epics/searchProducts.epic"
-import { getProductResults, SEND_PRODUCT_RESULTS } from "../../redux/actions"
+import {
+  getProductResults,
+  PRODUCT_ACTIONS
+} from "../../redux/actions/product.actions"
 
 // this creates an action observable, allowing us to fake an action that our epic can hear
 const action$ = ActionsObservable.of(getProductResults("socks"))
@@ -19,7 +22,7 @@ describe("getProductResults Epic", () => {
       }) // this will be replacing our ajax call!
 
     searchProducts(action$, null, { ajax }).subscribe(result => {
-      expect(result.type).toBe(SEND_PRODUCT_RESULTS)
+      expect(result.type).toBe(PRODUCT_ACTIONS.SEND_PRODUCT_RESULTS)
       expect(result.payload).toEqual(["socks", "mocks"])
 
       done()
